% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{All_header}
\alias{G}
\alias{M}
\alias{R}
\alias{example_data}
\alias{pheno}
\title{This is the data for examples}
\usage{
data(example_data)
}
\description{
\itemize{
  \item All_header. ID for all samples that are available in at least one type of data
  \item pheno. phenotype file. 1st column is ID, 2nd column is disease status, 3rd column is age, 4th column is gender
  \item G. genotypes for 3 genes. G[[i]] is for the ith genes. In each G[[i]], the 1st column is ID and the rest columns are genotypes
  \item M. methylated sites 3 genes. M[[i]] is for the ith genes. In each M[[i]], the 1st column is ID and the rest columns are methylated sites
  \item R. RNA expression for 3 genes. R[[i]] is for the ith genes. In each R[[i]], the 1st column is ID and the rest columns (usually one column) are RNA expression
}
}

